fix a bug with experience per level calculation

+++ src/player.cpp
@@ -98,7 +98,7 @@ namespace castlecrawl
         else
         {
             int exp = (t_level * 200);
-            exp += static_cast<int>(sqrt(static_cast<float>(exp))) * (t_level * 10);
+            exp += static_cast<int>(sqrt(static_cast<float>(exp)) * static_cast<float>(t_level * 10));
             return exp;
         }
     }
